==> 2_data_split.ipynb <==
# 2. Data Split

Now that we have filtered out outliers and have made sanity checks on the dataset, we can split the dataset into train, test and validation.

import pandas as pd
import numpy as np
from sklearn.model_selection import StratifiedGroupKFold, GroupShuffleSplit

df = pd.read_csv("cleaned_trajectories.csv", index_col=False)

Dataset details:

def dataset_details(df: pd.DataFrame) -> None:
    print(f"Number of unique aircraft: {df['flight_id'].nunique()}")
    print(f"Data length: {len(df)}")
    print(f"Data length by 30s windows: {len(df) // 30}")
    print(f"Data length by 60s windows: {len(df) // 60}")
    print(f"Data length by 120s windows: {len(df) // 120}")

    print("\n" + 15 * "-")
    print("Dataset info:")
    print(df.info(show_counts=True))

    print("\n" + 15 * "-")
    print("Dataset na percentages:")
    na_pct = df.isna().mean().sort_values(ascending=False) * 100
    display(na_pct)
    print("Any na values in dataset:", df.isna().values.any())

    print("\n" + 15 * "-")
    print(f"Unique flight ids: {df['flight_id'].nunique()}")

    print("\n" + 15 * "-")
    dupes = df.duplicated(subset=["flight_id", "timestamp"])
    print("Exact duplicate rows:", dupes.sum())

    print("\n" + 15 * "-")
    print(
        f"Number of 7700 squawks: {df['is_7700'].sum()}, percent: {df['is_7700'].sum()/len(df)*100:.2f}%"
    )
    print(
        f"Number of flights that have the squawk: {df[df['is_7700']].flight_id.nunique()}, percent: {df[df['is_7700']].flight_id.nunique()/df['flight_id'].nunique()*100:.2f}%"
    )


dataset_details(df)

## Splitting options
- Flight Group
- Individual points (after window transformation)

Splitting by individual points (row-wise) introduces the risk of data-leakage, does not generalise to real-world scenarios and breaks the temporal correlation as adjacent rows are split apart.


On the other side, the split by flight group (**flight_id**) has zero risk in data-leakage, is conservative and realistic as the model can be deployed for future aircraft and preserves temporal correlation within each split (windows keep their chronology).

Moreover, for sequence taks, group-wise splitting is the standard.

The following split will be used:
- Train: 70% (approx. 562 flights)
- Validation: 15% (approx. 120 flights)
- Test: 15% (approx. 121 flights)

As seen in the dataset information, every single aircraft has at least **one 7700 row** (29% of all rows/windows), but the *ratio per flight* still varies.  

To address this matter, we will:
1. Compute per-flight emergency ratio;
2. Define emergency-rate bins for stratification
- We will use buckets which become a coarse "class" for emergency-share, so that when we split we preserve the overall distribution of **low-, mid-, and high-emergency flights**.
3. Split the data by ensuring all windows of any given **flight_id** stay together.

flights = (
    df.groupby("flight_id")["is_7700"]
    .mean()  # emergency share for that flight
    .reset_index()
    .rename(columns={"is_7700": "emg_ratio"})
)
flights.head(2)

Bin the per-flight emergency ratio into 4 buckets

bins = pd.cut(
    flights["emg_ratio"],
    bins=[0, 0.05, 0.5, 0.95, 1],
    labels=False,
    include_lowest=True,
)
bins

Stratified split by those bins while repsecting the flight_id

sgkf = StratifiedGroupKFold(
    n_splits=5, shuffle=True, random_state=25
)  # 4/5 flights in train
train_idx, temp_idx = next(sgkf.split(X=flights, y=bins, groups=flights["flight_id"]))

train_flights = flights.iloc[train_idx]["flight_id"]
temp_flights = flights.iloc[temp_idx]["flight_id"]

gss = GroupShuffleSplit(test_size=0.5, n_splits=1, random_state=25)
val_idx, test_idx = next(gss.split(temp_flights, groups=temp_flights))
val_flights = temp_flights.iloc[val_idx]
test_flights = temp_flights.iloc[test_idx]

train_df = df[df["flight_id"].isin(train_flights)]
val_df = df[df["flight_id"].isin(val_flights)]
test_df = df[df["flight_id"].isin(test_flights)]

val_flights

print(
    f"Split details: (percentages of flights - by flight_id): Train {len(train_flights)/len(flights):.4f}, Val {len(val_flights)/len(flights):.4f}, Test {len(test_flights)/len(flights):.4f}"
)
print(
    f"Split details: (percentage of split - by rows): Train {len(train_df)/len(df):.4f}, Val {len(val_df)/len(df):.4f}, Test {len(test_df)/len(df):.4f}"
)
print(
    f"Numer of train flights: {len(train_flights)}, share of emergency rows: {train_df['is_7700'].sum()/len(train_df):.4f}"
)
print(
    f"Numer of val flights: {len(val_flights)}, share of emergency rows: {val_df['is_7700'].sum()/len(val_df):.4f}"
)
print(
    f"Numer of test flights: {len(test_flights)}, share of emergency rows: {test_df['is_7700'].sum()/len(test_df):.4f}"
)

Check for the correctness of the split

check_train_flights = train_df["flight_id"].unique()
check_val_flights = val_df["flight_id"].unique()
check_test_flights = test_df["flight_id"].unique()

for ff in check_train_flights:
    if ff in check_val_flights:
        print(f"Train flight {ff} is also in val set")
    if ff in check_test_flights:
        print(f"Train flight {ff} is also in test set")

for ff in check_val_flights:
    if ff in check_train_flights:
        print(f"Val flight {ff} is also in train set")
    if ff in check_test_flights:
        print(f"Val flight {ff} is also in test set")

data_split_folder = "data_splits/"
train_df.to_csv(data_split_folder + "train.csv", index=False)
val_df.to_csv(data_split_folder + "val.csv", index=False)
test_df.to_csv(data_split_folder + "test.csv", index=False)
print(f"Train, val, and test sets saved to {data_split_folder} folder.")
