==> 3_data_preprocessing_baseline.ipynb <==
# 3. Data Preprocessing/Transformation

Now that we have the data correct data splits, we will transform the data.

import numpy as np
import pandas as pd
from sklearn.preprocessing import StandardScaler
import joblib

data_splits_folder = "data_splits/"
train_df = pd.read_csv(data_splits_folder + "train.csv", parse_dates=["timestamp"])
test_df = pd.read_csv(data_splits_folder + "test.csv", parse_dates=["timestamp"])
val_df = pd.read_csv(data_splits_folder + "val.csv", parse_dates=["timestamp"])

def dataset_details(df: pd.DataFrame) -> None:
    print(f"Number of unique aircraft: {df['flight_id'].nunique()}")
    print(f"Data length: {len(df)}")
    print(f"Data length by 30s windows: {len(df) // 30}")
    print(f"Data length by 60s windows: {len(df) // 60}")
    print(f"Data length by 120s windows: {len(df) // 120}")

    print("\n" + 15 * "-")
    print("Dataset info:")
    print(df.info(show_counts=True))

    print("\n" + 15 * "-")
    print("Dataset na percentages:")
    na_pct = df.isna().mean().sort_values(ascending=False) * 100
    display(na_pct)
    print("Any na values in dataset:", df.isna().values.any())

    print("\n" + 15 * "-")
    print(f"Unique flight ids: {df['flight_id'].nunique()}")

    print("\n" + 15 * "-")
    dupes = df.duplicated(subset=["flight_id", "timestamp"])
    print("Exact duplicate rows:", dupes.sum())

    print("\n" + 15 * "-")
    print(
        f"Number of 7700 squawks: {df['is_7700'].sum()}, percent: {df['is_7700'].sum()/len(df)*100:.2f}%"
    )
    print(
        f"Number of flights that have the squawk: {df[df['is_7700']].flight_id.nunique()}, percent: {df[df['is_7700']].flight_id.nunique()/df['flight_id'].nunique()*100:.2f}%"
    )


dataset_details(train_df)

dataset_details(test_df)

dataset_details(val_df)

train_df.head()

# The dataset is cleaned of NaNs but we can use the following as future proof
# num_cols_to_fill = ["groundspeed", "track", "vertical_rate"]
# for _df in (train_df, val_df, test_df):
#     _df.sort_values("timestamp", inplace=True)
#     _df[num_cols_to_fill] = (_df.set_index("timestamp")[num_cols_to_fill]
#                                .interpolate("time")
#                                .fillna(method="ffill")
#                                .reset_index(drop=True))

## 1. Angle encoding
All circular features (latitude, longitude, track) are replaced by sine & cosine pairs. This turns each degree value into two continuous components (_sin, _cos), which avoids the discontinuity at 0°/360° and lets the model learn smooth angular patterns.


def add_angle_features(df, col_deg, prefix):
    rad = np.deg2rad(df[col_deg])
    df[f"{prefix}_sin"] = np.sin(rad)
    df[f"{prefix}_cos"] = np.cos(rad)
    df.drop(columns=[col_deg], inplace=True)


for _df in (train_df, val_df, test_df):
    add_angle_features(_df, "latitude", "lat")
    add_angle_features(_df, "longitude", "lon")
    add_angle_features(_df, "track", "trk")

train_df.head()

## 2. Turn-rate feature
A crude d(track)/dt is computed by differencing the trk_cos column within each flight.

This new column is then standardized using a second StandardScaler (again trained on train, then applied to val/test).

for _df in (train_df, val_df, test_df):
    _df["turn_rate"] = _df.groupby("flight_id")["trk_cos"].diff().fillna(0)

train_df.head()

## 3. Global z-score scaling
A StandardScaler is fitted only on the training split’s nine core features (altitude, groundspeed, vertical_rate, and the three sin/cos pairs); as well as the added turn rate.

- The training data are fit_transformed.

- Validation and test are then transformed with the same scaler.

- The fitted scaler is saved to scaler.pkl for later reuse.

feature_cols = [
    "altitude",
    "groundspeed",
    "vertical_rate",
    "lat_sin",
    "lat_cos",
    "lon_sin",
    "lon_cos",
    "trk_sin",
    "trk_cos",
    "turn_rate",
]

scaler = StandardScaler()
train_df[feature_cols] = scaler.fit_transform(train_df[feature_cols])
val_df[feature_cols] = scaler.transform(val_df[feature_cols])
test_df[feature_cols] = scaler.transform(test_df[feature_cols])

data_processing_folder = "baseline_data_processing/"
joblib.dump(scaler, data_processing_folder + "scaler.pkl")

train_df.head()



## 4. Sliding-window tensorization
For each flight:

- Extract its feature matrix (T rows × F features).

- Slide a window of length win_len with step stride along the time axis.

- For each window, stack its win_len × F values into a sample, and set its label to 1 if any is_7700 flag occurs in that window (otherwise 0).

- Return three arrays per split:

> X with shape (N_windows, win_len, n_features)

> y with shape (N_windows,)

Stride example:

win_len =  30, stride=10
- Window 1: rows **0–29**  
- Window 2: rows **10–39**  
- Window 3: rows **20–49**  

def build_windows(df, win_len, stride, feature_cols):
    data, labels = [], []
    for _, fl in df.groupby("flight_id"):
        X = fl[feature_cols].to_numpy(dtype=np.float32)
        y = fl["is_7700"].to_numpy(dtype=np.int8)  # chdanged from bool to int
        for i in range(0, len(X) - win_len + 1, stride):
            data.append(X[i : i + win_len])
            labels.append(int(y[i : i + win_len].any()))  # 1 if ANY emergency in window
    return np.stack(data), np.array(labels)

WIN = 30  # seconds, can be tuned
STRIDE = 10

X_train, y_train = build_windows(train_df, WIN, STRIDE, feature_cols)
X_val, y_val = build_windows(val_df, WIN, STRIDE, feature_cols)
X_test, y_test = build_windows(test_df, WIN, STRIDE, feature_cols)

X_train.shape

The shape confirms the type of data - multivariate time-series.
It means that we have:
- 332142 samples (independent)
- ordered sequence of 30 -1Hz frames
- 10 parallel channels of time series

Save the dataset:

payload = {
    "X_train": X_train,
    "y_train": y_train,
    "X_val": X_val,
    "y_val": y_val,
    "X_test": X_test,
    "y_test": y_test,
    "scaler": scaler,
    "train_df": train_df,
    "val_df": val_df,
    "test_df": test_df,
    "feature_cols": feature_cols,
    "win_len": WIN,
    "stride": STRIDE,
}

joblib.dump(payload, data_processing_folder + "data_payload_30s.pkl")
